# public/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "VETRA API - Platform Kesehatan Hewan Peliharaan",
    "description": "Dokumentasi Lengkap API untuk VETRA - Platform Booking Klinik, Konsultasi Dokter, Chatting, dan AI Chatbot",
    "version": "1.0.0",
    "contact": {
      "name": "Vetra API Support",
      "email": "support@vetra.com"
    }
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8000/api",
      "description": "Development Server"
    }
  ],
...
  "paths": {
...
    "/clinics": {
      "get": {
        "tags": ["Public Content"],
        "summary": "Mendapatkan daftar klinik",
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } }
        ],
        "responses": {
          "200": { "description": "List data klinik" }
        }
      }
    },
    "/clinics/{id}": {
      "get": {
        "tags": ["Public Content"],
        "summary": "Mendapatkan detail klinik",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": {
          "200": { "description": "Detail profile klinik" }
        }
      }
    },
...
    "/clinics/{id}/doctors": {
      "get": {
        "tags": ["Public Content"],
        "summary": "Mendapatkan daftar dokter dalam klinik tersebut",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": {
          "200": { "description": "Daftar dokter di klinik" }
        }
      }
    },
...
    "/clinic/dashboard": {
      "get": {
        "tags": ["Clinic Panel"],
        "summary": "Dashboard statistik klinik",
        "security": [{"bearerAuth": []}],
        "responses": {
          "200": { "description": "Statistik global klinik" }
        }
      }
    },
    "/clinic/bookings": {
      "get": {
        "tags": ["Clinic Panel"],
        "summary": "Mengambil data booking klinik",
        "security": [{"bearerAuth": []}],
        "responses": {
          "200": { "description": "List data booking klinik" }
        }
      }
    },
    "/clinic/doctors": {
      "get": {
        "tags": ["Clinic Panel"],
        "summary": "Mengambil dokter klinik tersebut",
        "security": [{"bearerAuth": []}],
        "responses": {
          "200": { "description": "List dokter klinik" }
        }
      },
      "post": {
        "tags": ["Clinic Panel"],
        "summary": "Menambahkan dokter ke klinik",
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "email", "password"],
                "properties": {
                  "name": { "type": "string", "example": "Dr. Sarah" },
                  "email": { "type": "string", "format": "email", "example": "sarah@vetra.id" },
                  "password": { "type": "string", "example": "password123" },
                  "phone": { "type": "string", "example": "08123445" },
                  "spesialis": { "type": "string", "example": "Dokter Umum" },
                  "experience_years": { "type": "integer", "example": 3 }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Dokter berhasil ditambahkan" }
        }
      }
    },
    "/clinic/doctors/{id}": {
      "put": {
        "tags": ["Clinic Panel"],
        "summary": "Update data dokter oleh klinik",
        "security": [{"bearerAuth": []}],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "spesialis": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Data dokter diperbarui" }
        }
      },
      "delete": {
        "tags": ["Clinic Panel"],
        "summary": "Hapus dokter dari klinik",
        "security": [{"bearerAuth": []}],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": {
          "200": { "description": "Dokter terhapus" }
        }
      }
    },
...
    "/v2/clinics": {
      "get": {
        "tags": ["API Key Integration (v2)"],
        "summary": "List klinik (API Key)",
        "security": [{"apiKeyAuth": []}],
        "responses": {
          "200": { "description": "List data klinik" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Autentikasi standar JWT. Masukkan token Anda (Tanpa menulis 'Bearer ')"
      },
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Autentikasi integrasi API Key statis. Kirimkan token API Key Anda pada header X-API-Key"
      }
    },
...
  }
}